Change behaviour to error on empty HardhatConfig
If a user fails to export from the HardhatConfig.{js,ts} file or exports an empty object, we now throw an error.

This means updating some of our tests cases which have been using the `{}` object as the base case. The docs (including plugin docs) have been updated to reflect that `module.exports = {}` will not work when copied and pasted.

Relates to #1490